Move test helper attestation aggregation methods from the jmh-benchmark module to test-support
