+++ eth-benchmark-tests/src/jmh/java/tech/pegasys/artemis/benchmarks/gen/Utils.java
@@ -13,57 +13,11 @@
 
 package tech.pegasys.artemis.benchmarks.gen;
 
-import com.google.common.base.Preconditions;
-import java.util.Collection;
 import java.util.Iterator;
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
-import tech.pegasys.artemis.datastructures.operations.Attestation;
-import tech.pegasys.artemis.util.SSZTypes.Bitlist;
-import tech.pegasys.artemis.util.bls.BLSAggregate;
-import tech.pegasys.artemis.util.bls.BLSSignature;
-import tech.pegasys.artemis.util.config.Constants;
 
 public class Utils {
-  /**
-   * Groups passed attestations by their {@link
-   * tech.pegasys.artemis.datastructures.operations.AttestationData} and aggregates attestations in
-   * every group to a single {@link Attestation}
-   *
-   * @return a list of aggregated {@link Attestation}s with distinct {@link
-   *     tech.pegasys.artemis.datastructures.operations.AttestationData}
-   */
-  public static List<Attestation> groupAndAggregateAttestations(List<Attestation> srcAttestations) {
-    Collection<List<Attestation>> groupedAtt =
-        srcAttestations.stream().collect(Collectors.groupingBy(Attestation::getData)).values();
-    return groupedAtt.stream().map(Utils::aggregateAttestations).collect(Collectors.toList());
-  }
-
-  /**
-   * Aggregates passed attestations
-   *
-   * @param srcAttestations attestations which should have the same {@link Attestation#getData()}
-   */
-  public static Attestation aggregateAttestations(List<Attestation> srcAttestations) {
-    Preconditions.checkArgument(!srcAttestations.isEmpty(), "Expected at least one attestation");
-
-    int targetBitlistSize =
-        srcAttestations.stream()
-            .mapToInt(a -> a.getAggregation_bits().getCurrentSize())
-            .max()
-            .getAsInt();
-    Bitlist targetBitlist = new Bitlist(targetBitlistSize, Constants.MAX_VALIDATORS_PER_COMMITTEE);
-    srcAttestations.forEach(a -> targetBitlist.setAllBits(a.getAggregation_bits()));
-    BLSSignature targetSig =
-        BLSAggregate.bls_aggregate_signatures(
-            srcAttestations.stream()
-                .map(Attestation::getAggregate_signature)
-                .collect(Collectors.toList()));
-
-    return new Attestation(targetBitlist, srcAttestations.get(0).getData(), targetSig);
-  }
 
   static <T> Iterator<T> fromSupplier(Supplier<T> supplier) {
     return new Iterator<T>() {
